Factored out NotificationManager
