+++ src/components/Notifications/index.ts
@@ -16,3 +16,4 @@
  */
 
 export * from "./Notifications";
+export * from "./NotificationManager";
